+++ .vscode/settings.json
@@ -89,6 +89,7 @@
     "userid",
     "vitepress",
     "withs",
+    "xmlparser",
     "zcube"
   ]
 }
