--- runtimes/nodejs/scripts/create-internal-package.js
@@ -1,5 +1,0 @@
-const { createCloudFunctionDeclarationPackage } = require("../dist/support/init");
-
-createCloudFunctionDeclarationPackage()
-
-process.exit(0)
