+++ runtimes/nodejs/src/handler/router.ts
@@ -31,7 +31,7 @@ export const router = Router()
 
 router.post('/proxy/:policy', handleDatabaseProxy)
 router.get('/typing/package', handlePackageTypings)
-router.post('/_/healthz', (_req, res) => res.status(200).send('ok'))
+router.get('/_/healthz', (_req, res) => res.status(200).send('ok'))
 router.post('/_/publish/functions', (_req, res) => res.status(400).send('TODO'))
 router.post('/_/publish/policies', handlePublishPolicies)
